chore: move references to plugin.json to new file to prevent jest tests from failing
