+++ src/services/plugin.ts
@@ -1,0 +1,9 @@
+import pluginJson from '../plugin.json';
+
+export const PLUGIN_ID = pluginJson.id;
+export const PLUGIN_BASE_URL = `/a/${PLUGIN_ID}`;
+
+// Prefixes the route with the base URL of the plugin
+export function prefixRoute(route: string): string {
+  return `${PLUGIN_BASE_URL}/${route}`;
+}
